requirements.txt
streamlit
pandas
python-dotenv
azure-identity
openai